requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
google-genai>=1.0.0
openai>=1.0.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.23.0
